requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
groq==0.4.1
pillow==10.1.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0